requirements.txt
python-telegram-bot==20.7
Flask==3.0.2
gspread==6.0.2
oauth2client==4.1.3
